+++ FlexiGlob.Zip/AssemblyAttributes.cs
@@ -1,4 +1,3 @@
 ﻿using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("FlexiGlob.UnitTests")]
-[assembly: InternalsVisibleTo("FlexiGlob.Zip")]