+++ FlexiGlob/ITemporaryDirectory.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace FlexiGlob
+{
+    public interface ITemporaryDirectory : IDisposable
+    {
+        string FullPath { get; }
+        void CleanUp();
+    }
+}